Update campayn.php
I corrected the Mixed Media issue. 
It happened while viewing from an https (SSL) domain. 
It had to do with the form posting to "http://campayn.com/contacts/signup_form_add_contact/"

So in the code I rewrite it to https, and also provided a way to rewrite it using the protocol used utilizing the existing if statement.